The Utilities !
ShaderMotion supports really brought more pain than anything.
One of these pain was to make some conversion script that
"just works", meaning that people would have to only right-click
somewhere, select the save path and... DONE !

The script is now created. It still depends on VRM and I don't
really know how I should handle this.
Still, with the right packages installed, it works !

And no, you don't need an external installation of ShaderMotion,
this includes a special version with it.

That special version depends on the shaders I created for
the 3D Model Loader though, hence why I started to split the
3D Model Loader into multiple sub-packages.

Which brought dependency hell ! And UPM SUCK BALLS in that part.
UPM cannot deal with Zip files, nor can it deal easily with Git
dependencies.
It can pull a package from Git but it cannot pull other dependencies
from Git.

I'll try the UPM route for bringing the VRM dependency.
But if that doesn't work, I'll just rely on the good will of the
user installing this package.

Signed-off-by: Voyage <voyage@miouyouyou.fr>